+++ ast/expr-operator-unary.go
@@ -4,7 +4,7 @@ var _ Expr = ExprOperatorUnary{}
 
 type ExprOperatorUnary struct {
 	Operator OperatorUnary
-	Expr     *Expr
+	Expr     Expr
 
 	MetaInfo *MetaExpr
 }
@@ -13,6 +13,16 @@ func (e ExprOperatorUnary) Meta() *MetaExpr {
 	return e.MetaInfo
 }
 
+func MakeExprOperatorUnary(operator OperatorUnary, expr Expr, source *Source) *ExprOperatorUnary {
+	return &ExprOperatorUnary{
+		Operator: operator,
+		Expr:     expr,
+		MetaInfo: &MetaExpr{
+			Source: source,
+		},
+	}
+}
+
 func (e ExprOperatorUnary) EnumerateNestedDecls(enumerate func(interface{}, []Decl)) {
-	(*e.Expr).EnumerateNestedDecls(enumerate)
+	e.Expr.EnumerateNestedDecls(enumerate)
 }