+++ parser/parse-expr-unary.go
@@ -4,6 +4,16 @@ import (
 	"github.com/vknabel/lithia/ast"
 )
 
-func (fp *FileParser) ParseUnaryExpr() (*ast.ExprInt, []SyntaxError) {
-	return nil, []SyntaxError{fp.SyntaxErrorf("unimplemented")}
+func (fp *FileParser) ParseUnaryExpr() (*ast.ExprOperatorUnary, []SyntaxError) {
+	if fp.Node.NamedChildCount() != 1 {
+		return nil, []SyntaxError{fp.SyntaxErrorf("expected one child, got %d", fp.Node.NamedChildCount())}
+	}
+	operator := fp.Node.ChildByFieldName("operator").Content(fp.Source)
+
+	exprP := fp.SameScopeChildParser(fp.Node.NamedChild(0))
+	expr, errs := exprP.ParseExpression()
+	if len(errs) > 0 {
+		return nil, errs
+	}
+	return ast.MakeExprOperatorUnary(ast.OperatorUnary(operator), expr, fp.AstSource()), nil
 }